rolled back changes to StoreText() at Jan 27 2009; fixed a bug where word1.\f resulted in word1. being treated as the word instead of word1, and . creeping into the KB as a result; added doc-healing code (for m_key and m_adaption to guarantee no punctuation in either) to RedoStorage(), as a side effect for Restore KB operation.

+++ source/Adapt_ItDoc.cpp
@@ -6643,7 +6643,8 @@ m:				if (bStarted && (wxUint32)(pPunctEnd - pPunctStart) > 0 && pPunctEnd == pt
 					// two bytes long, so setting numChars to the pointer difference will
 					// double the correct value; we have to therefore divide by sizeof(wxChar)
 					// to get numChars right in regular and unicode apps
-					int numChars = (int)(pPunctEnd - pPunctStart) / sizeof(wxChar);
+					//int numChars = (int)(pPunctEnd - pPunctStart) / sizeof(wxChar); //bad
+					int numChars = ((int)pPunctEnd - (int)pPunctStart) / sizeof(wxChar);
 					wxString finals(pPunctStart,numChars);
 					followPunct = finals;
 				}
